+++ src/commands/music/mod.rs
@@ -1,11 +1,17 @@
 mod disconnect;
+mod jam;
 mod join;
 mod lyrics;
 mod playing;
+mod shuffle;
 mod stop;
+mod volume;
 
 pub use disconnect::*;
+pub use jam::*;
 pub use join::*;
 pub use lyrics::*;
 pub use playing::*;
+pub use shuffle::*;
 pub use stop::*;
+pub use volume::*;